chore: add Zod compatibility version
