+++ examples/openapi-ts-nuxt/package.json
@@ -16,7 +16,7 @@
     "nuxt": "3.14.1592",
     "vue": "3.5.13",
     "vue-router": "4.5.0",
-    "zod": "3.23.8"
+    "zod": "3.25.0"
   },
   "devDependencies": {
     "vite": "6.2.7"
