+++ examples/openapi-ts-sample/package.json
@@ -18,7 +18,7 @@
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "valibot": "1.1.0",
-    "zod": "3.23.8"
+    "zod": "3.25.0"
   },
   "devDependencies": {
     "@config/vite-base": "workspace:*",
