+++ packages/openapi-ts/package.json
@@ -94,7 +94,8 @@
     "color-support": "1.1.3",
     "commander": "13.0.0",
     "handlebars": "4.7.8",
-    "open": "10.1.2"
+    "open": "10.1.2",
+    "semver": "7.7.2"
   },
   "peerDependencies": {
     "typescript": "^5.5.3"
@@ -104,6 +105,7 @@
     "@types/bun": "1.2.19",
     "@types/cross-spawn": "6.0.6",
     "@types/express": "4.17.21",
+    "@types/semver": "7.7.0",
     "axios": "1.8.2",
     "cross-spawn": "7.0.5",
     "eslint": "9.17.0",
