+++ packages/openapi-ts/src/config/utils/config.ts
