+++ packages/openapi-ts/src/plugins/zod/config.ts
@@ -1,7 +1,13 @@
+import colors from 'ansi-colors';
+
 import { definePluginConfig, mappers } from '../shared/utils/config';
 import { api } from './api';
 import { handler } from './plugin';
 import type { ZodPlugin } from './types';
+
+type CompatibilityVersion = NonNullable<
+  ZodPlugin['Config']['config']['compatibilityVersion']
+>;
 
 export const defaultConfig: ZodPlugin['Config'] = {
   api,
@@ -15,6 +21,50 @@ export const defaultConfig: ZodPlugin['Config'] = {
   name: 'zod',
   output: 'zod',
   resolveConfig: (plugin, context) => {
+    const packageName = 'zod';
+    const version = context.package.getVersion(packageName);
+
+    const inferCompatibleVersion = (): CompatibilityVersion => {
+      if (version && (version.major === 4 || version.major === 3)) {
+        return version.major;
+      }
+
+      // default compatibility version
+      return 4;
+    };
+
+    const ensureCompatibleVersion = (
+      compatibilityVersion: CompatibilityVersion | undefined,
+    ): CompatibilityVersion => {
+      if (!compatibilityVersion) {
+        return inferCompatibleVersion();
+      }
+
+      if (!version) {
+        return compatibilityVersion;
+      }
+
+      if (
+        compatibilityVersion === 4 ||
+        compatibilityVersion === 3 ||
+        compatibilityVersion === 'mini'
+      ) {
+        if (!context.package.satisfies(version, '>=3.25.0 <5.0.0')) {
+          const compatibleVersion = inferCompatibleVersion();
+          console.warn(
+            `🔌 ${colors.yellow('Warning:')} Installed ${colors.cyan(packageName)} ${colors.cyan(`v${version.version}`)} does not support compatibility version ${colors.yellow(String(compatibilityVersion))}, using ${colors.yellow(String(compatibleVersion))}.`,
+          );
+          return compatibleVersion;
+        }
+      }
+
+      return compatibilityVersion;
+    };
+
+    plugin.config.compatibilityVersion = ensureCompatibleVersion(
+      plugin.config.compatibilityVersion,
+    );
+
     plugin.config.dates = context.valueToObject({
       defaultValue: {
         offset: false,
