+++ packages/openapi-ts/src/plugins/zod/shared/types.d.ts
@@ -1,0 +1,6 @@
+import type ts from 'typescript';
+
+export type ZodSchema = {
+  expression: ts.Expression;
+  typeName?: string;
+};
